==> Books-Sales-Recommendation-Algorithm-main/Codes/2-Data_Cleaning.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_feather('raw_data_unclean.feather').drop(['formato','language','editora'],axis=1)

df.isnull().sum()

df.dtypes

df = df.dropna()

strings = df2['sales']
newsales = []
for e in strings:
    try:    
        try:
            k = e.replace('Novo  |  ','')
            try:
                newsales.append(float(k.replace('vendidos','')))
            except:
                newsales.append(float(k.replace('vendido','')))
        
        except:
                k = e.replace('Usado  |  ','')
                try:
                    newsales.append(float(k.replace('vendidos','')))
                except:
                    newsales.append(float(k.replace('vendido','')))
            
           
    except:
        newsales.append('null')
        continue
               
            
        
newsales = pd.DataFrame(newsales)

index_names = newsales[newsales[0] == 'null'].index
newsales = newsales.drop(index_names,axis=0)

df['sales'] = newsales[0]
df = df.dropna().reset_index().drop('index',axis=1)

df.isnull().sum()

df.dtypes

df.to_csv('raw_data_nlabels.csv')
