==> Books-Sales-Recommendation-Algorithm-main/Codes/4.Ensamble.ipynb <==
import pandas as pd
import numpy as np



from sklearn.pipeline import make_pipeline
from sklearn.preprocessing import MaxAbsScaler,StandardScaler
from scipy.sparse import hstack,csr_matrix
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import TfidfVectorizer 
from yellowbrick.classifier import ConfusionMatrix
from sklearn.metrics import average_precision_score , roc_auc_score

df = pd.read_csv('raw_data_wlabels.csv').drop('Unnamed: 0',axis=1).dropna()

y = df['y']
xdata = df.drop('y',axis=1)

data_train,data_val,ytrain,yval = train_test_split(xdata,
                                                    y,
                                                    test_size = 0.5,
                                                    random_state=0)

# Vetorizando as features de texto
title_train = data_train['titles']
title_val = data_val['titles']

title_vec = TfidfVectorizer(min_df = 2,ngram_range=(1,1))

title_vec3 = TfidfVectorizer(min_df = 4,ngram_range=(1,5))

title_bow_train = title_vec.fit_transform(title_train)
title_bow_val = title_vec.transform(title_val)

title_bow_train3 = title_vec3.fit_transform(title_train)
title_bow_val3 = title_vec3.transform(title_val)


autor_train = data_train['author']
autor_val = data_val['author']

autor_vec = TfidfVectorizer(min_df = 2,ngram_range=(1,1))
autor_vec = TfidfVectorizer(min_df = 4,ngram_range=(1,5))

autor_bow_train = autor_vec.fit_transform(autor_train)
autor_bow_val = autor_vec.transform(autor_val)

autor_bow_train3 = autor_vec3.fit_transform(autor_train)
autor_bow_val3 = autor_vec3.transform(autor_val)

mask_train = data_train.drop(['titles','author'],axis=1)
mask_val = data_val.drop(['titles','author'],axis=1)

xtrain_wvec = hstack([title_bow_train,autor_bow_train,mask_train])
xval_wvec = hstack([title_bow_val,autor_bow_val,mask_val])

xtrain_wvec3 = hstack([title_bow_train3,autor_bow_train3,mask_train])
xval_wvec3 = hstack([title_bow_val3,autor_bow_val3,mask_val])


# Regressão Logistica

xtrain_wvec2 = csr_matrix(xtrain_wvec)
xval_wvec2 = csr_matrix(xval_wvec)

from sklearn.linear_model import LogisticRegression
lr = make_pipeline(MaxAbsScaler(),LogisticRegression(C=30,n_jobs=6,random_state=0))
lr.fit(xtrain_wvec2,ytrain)

p_lr = lr.predict_proba(xval_wvec2)[:,1]
print(f'AVP: {average_precision_score(yval,p_lr)}')    
print(f'ROC: {roc_auc_score(yval,p_lr)}')

# Random Forest


from sklearn.ensemble import RandomForestClassifier
rdf = RandomForestClassifier(n_estimators = 1000,random_state=0,min_samples_leaf=3,class_weight = 'balanced', n_jobs=6)
rdf = rdf.fit(csr_matrix(xtrain_wvec),ytrain)

p_rd = rdf.predict_proba(csr_matrix(xval_wvec))[:,1]
print(f'AVP: {average_precision_score(yval,p_rd)}')    
print(f'ROC: {roc_auc_score(yval,p_rd)}')

# Árvore de Decisão

from sklearn.tree import DecisionTreeClassifier
dt = DecisionTreeClassifier(random_state = 0,max_depth = 4,class_weight = "balanced")
dt.fit(csr_matrix(xtrain_wvec),ytrain)

p_dt = dt.predict_proba(csr_matrix(xval_wvec))[:,1]
print(f'AVP: {average_precision_score(yval,p_dt)}')    
print(f'ROC: {roc_auc_score(yval,p_dt)}')

# LGBM

best = [0.0017606961005459585,10,2,0.7976031342753723,0.23056440691282926,906,4,5]

from lightgbm import LGBMClassifier
LGBM =  LGBMClassifier(learning_rate=best[0], num_leaves= 2 ** best[1], max_depth=best[1],
                         min_child_samples=best[2], subsample=best[3],
                         colsample_bytree=best[4], bagging_freq = 1, n_estimators=best[5], 
                         random_state = 0, class_weight ='balanced',n_jobs=6)
LGBM.fit(csr_matrix(xtrain_wvec3),ytrain)

p_lgb = LGBM.predict_proba(csr_matrix(xval_wvec3))[:,1]

p_lgb = LGBM.predict_proba(csr_matrix(xval_wvec3))[:,1]
print(f'AVP: {average_precision_score(yval,p_lgb)}')    
print(f'ROC: {roc_auc_score(yval,p_lgb)}')

# Montando o Ensamble

pd.DataFrame({'LR': p_lr,'RDF':p_rd,'DCT':p_dt,'LGBM':p_lgb}).corr()

p = (0.25*p_rd + 0.25*p_lr + 0.5*p_lgb)

print(f'AVG : {average_precision_score(yval,p)} \nROC : {roc_auc_score(yval,p)}')

# Salvando os Modelos

import joblib as jb
jb.dump(lr,'logistic_reg_2021_05_20.pkl')
jb.dump(rdf,'random_for_2021_05_20.pkl')
jb.dump(LGBM,'LGBM_2021_05_20.pkl')

jb.dump(title_vec,'title_vectorizer_2021_05_20.pkl')
jb.dump(title_vec3,'title_vectorizer3_2021_05_20.pkl')
jb.dump(autor_vec,'autor_vectorizer_2021_05_20.pkl')
